refactor(html-spec)!: migrate spec-generator into html-spec and run via native TypeScript
Move all @markuplint/spec-generator source files into
html-spec/generator/ and delete the spec-generator package.

- Rewrite import extensions from .js to .ts for Node 24 native
  TypeScript execution (type stripping enabled by default)
- Replace build.mjs with build.ts using local imports
- Replace gen:prettier with gen:format using oxfmt
- Add spec-generator runtime deps as html-spec devDependencies

BREAKING CHANGE: @markuplint/spec-generator package is removed.
Its functionality is now internal to @markuplint/html-spec.
