+++ packages/@markuplint/html-spec/generator/read-json.ts
