+++ packages/@markuplint/html-spec/generator/utils.ts
