--- packages/@markuplint/spec-generator/.gitignore
@@ -1,1 +1,0 @@
-__*.json
