--- packages/@markuplint/spec-generator/.npmignore
@@ -1,2 +1,0 @@
-src
-tsconfig*
