--- packages/@markuplint/spec-generator/package.json
@@ -1,41 +1,0 @@
-{
-	"name": "@markuplint/spec-generator",
-	"version": "5.0.0-rc.1",
-	"description": "Generates @markuplint/html-spec",
-	"repository": "git@github.com:markuplint/markuplint.git",
-	"author": "Yusuke Hirao <yusukehirao@me.com>",
-	"license": "MIT",
-	"engines": {
-		"node": ">=22"
-	},
-	"type": "module",
-	"exports": {
-		".": {
-			"import": "./lib/index.js",
-			"types": "./lib/index.d.ts"
-		}
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"scripts": {
-		"build": "tsc --project tsconfig.build.json",
-		"dev": "tsc --watch --project tsconfig.build.json",
-		"clean": "tsc --build --clean tsconfig.build.json"
-	},
-	"dependencies": {
-		"@types/cheerio": "1.0.0",
-		"ajv": "8.18.0",
-		"cheerio": "1.2.0",
-		"cli-progress": "3.12.0",
-		"fast-xml-parser": "5.5.9",
-		"glob": "13.0.6",
-		"jsonc-parser": "3.3.1"
-	},
-	"devDependencies": {
-		"@markuplint/ml-spec": "5.0.0-rc.1",
-		"@markuplint/test-tools": "5.0.0-rc.1",
-		"@types/cli-progress": "3.11.6",
-		"type-fest": "5.5.0"
-	}
-}
