--- packages/@markuplint/spec-generator/tsconfig.json
@@ -1,11 +1,0 @@
-{
-	"extends": "../../../tsconfig.json",
-	"compilerOptions": {
-		"composite": true
-	},
-	"references": [
-		{
-			"path": "../ml-spec"
-		}
-	]
-}
